// src/app/@data/models/DetalleDeuda.ts
export class DetalleDeuda {
    idEmpresa:        number;
    idTipoDocumento:  number;
    nroDocumento:     string;
    idPeriodo:        number;
    fechaVencimiento: Date;
    saldo:            string;
    isMinimo:         number;
    nombreMes:        string;
    mes:              number;
    disabled:           boolean;
}